+++ exeris-kernel-community/src/test/java/eu/exeris/kernel/community/bootstrap/CommunityPersistenceBootstrapIntegrationTest.java
@@ -189,7 +189,10 @@ class CommunityPersistenceBootstrapIntegrationTest {
                                             HttpConfig.DEFAULT_MAX_REQUEST_BODY_BYTES,
                                             false,
                                             HttpVersion.HTTP_1_1,
-                                            "127.0.0.1" + ":" + port
+                                            "127.0.0.1" + ":" + port,
+                                            HttpConfig.DEFAULT_MAX_HEADER_BLOCK_SIZE,
+                HttpConfig.DEFAULT_MAX_HEADER_LIST_SIZE,
+                HttpConfig.DEFAULT_MAX_STRING_LITERAL_SIZE
                                     );
                                     try (HttpClientEngine client = new CommunityHttpProvider().createClientEngine(clientConfig)) {
                                         client.start();